// src/app/services/models/book-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookDto {
  bookAsin?: string;
  bookAuthors?: Array<string>;
  bookEdition?: number;
  bookEntryInSeries?: number;
  bookId?: number;
  bookIsbn?: string;
  bookPages?: number;
  bookPublisherName?: string;
  bookRelease?: string;
  bookSeriesName?: string;
  bookSynopsis?: string;
  bookTitle?: string;
}