// src/app/services/models/author-dto.ts
/* tslint:disable */
/* eslint-disable */

/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AuthorDto {
  authorBooks?: Array<string>;
  authorFullName?: string;
  authorId?: number;
  authorName?: string;
  authorNumberOfBooks?: number;
  authorYearOfBirth?: number;
  authorYearOfDeath?: number;
}