// src/app/services/models/publisher-dto.ts
/* tslint:disable */
/* eslint-disable */

/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PublisherDto {
  publisherBooks?: Array<string>;
  publisherId?: number;
  publisherName?: string;
  publisherNumberOfBooks?: number;
  publisherOriginCity?: string;
  publisherOriginCountry?: string;
}